chore(actions): moved actions to their own files
