+++ src/actions/game-tied.js
@@ -1,0 +1,7 @@
+import { GAME_TIED } from './types'
+
+export default function gameTied() {
+  return {
+    type: GAME_TIED
+  }
+}
