+++ src/actions/game-won.js
@@ -1,0 +1,8 @@
+import { GAME_WON } from './types'
+
+export default function gameWon(player) {
+  return {
+    type: GAME_WON,
+    payload: player
+  }
+}
