+++ src/actions/player-move.js
@@ -1,0 +1,5 @@
+import { PLAYER_MOVE } from './types'
+
+export default function playerMove(coords) {
+  return { type: PLAYER_MOVE, payload: coords }
+}
