+++ src/actions/reset-game.js
@@ -1,0 +1,9 @@
+import { RESET_GAME } from './types'
+
+function resetGame() {
+  return {
+    type: RESET_GAME
+  }
+}
+
+export default resetGame
